Moved `utils` to the new home pioneered in https://github.com/Octav-Labs/sonarwatch-portfolio/pull/35 and adjusted `import`s
(I actually reformatted the changed files; this is why there are small changes)
