+++ packages/plugins/src/utils/octav/loggingUtils.ts
@@ -1,12 +1,12 @@
-/** 
+/**
  * A logging context is a record of key-value pairs that can be used to add additional context to a log message.
  * The only required key is `fn`, which is the name of the function that is logging the message.
  */
 export interface LoggingContext {
-    fn: string;
-    [key: string]: unknown;
-  }
-  
+  fn: string;
+  [key: string]: unknown;
+}
+
 /**
  * Logs a message with structured context information only when OCTAV_VERBOSE_LOGGING is enabled.
  * The below solution is not as fancy as supporting logger injection or eventing, but I really wanted
